refactor: split styles and related files about ScheduleModal
